backend/requirements.txt
flask
flask_pymongo
flask-cors
pymongo
requests
urllib3
python-dotenv